code with me [GeneralPashon] Camera, Camera3D, Matrix3, Matrix4, PerspectiveCamera, (Frustum, Vec3d, Vec3f, Vec3i, Quaternion, Shader)

+++ core/src/main/kotlin/pize/math/vecmath/matrix/Matrix3.kt
@@ -2,14 +2,14 @@ package pize.math.vecmath.matrix
 
 interface Matrix3 {
     companion object {
-        const val m00: Byte = 0
-        const val m10: Byte = 1
-        const val m20: Byte = 2
-        const val m01: Byte = 3
-        const val m11: Byte = 4
-        const val m21: Byte = 5
-        const val m02: Byte = 6
-        const val m12: Byte = 7
-        const val m22: Byte = 8
+        const val m00: Int = 0
+        const val m10: Int = 1
+        const val m20: Int = 2
+        const val m01: Int = 3
+        const val m11: Int = 4
+        const val m21: Int = 5
+        const val m02: Int = 6
+        const val m12: Int = 7
+        const val m22: Int = 8
     }
 }
