+++ core/src/main/kotlin/pize/math/vecmath/matrix/Matrix4.kt
@@ -2,21 +2,21 @@ package pize.math.vecmath.matrix
 
 interface Matrix4 {
     companion object {
-        const val m00: Byte = 0
-        const val m10: Byte = 1
-        const val m20: Byte = 2
-        const val m30: Byte = 3
-        const val m01: Byte = 4
-        const val m11: Byte = 5
-        const val m21: Byte = 6
-        const val m31: Byte = 7
-        const val m02: Byte = 8
-        const val m12: Byte = 9
-        const val m22: Byte = 10
-        const val m32: Byte = 11
-        const val m03: Byte = 12
-        const val m13: Byte = 13
-        const val m23: Byte = 14
-        const val m33: Byte = 15
+        const val m00: Int = 0
+        const val m10: Int = 1
+        const val m20: Int = 2
+        const val m30: Int = 3
+        const val m01: Int = 4
+        const val m11: Int = 5
+        const val m21: Int = 6
+        const val m31: Int = 7
+        const val m02: Int = 8
+        const val m12: Int = 9
+        const val m22: Int = 10
+        const val m32: Int = 11
+        const val m03: Int = 12
+        const val m13: Int = 13
+        const val m23: Int = 14
+        const val m33: Int = 15
     }
 }
